ПР14

+++ controllers/cards.js
@@ -1,22 +1,20 @@
 const Card = require('../models/card');
-const { ERROR_CODE, ERROR_NOT_FOUND_CODE, INTERNAL_ERROR } = require('../constants/networkStatuses');
+const ERROR_CODE = require('../errors/error_code');
+const ERROR_NOT_FOUND_CODE = require('../errors/error_not_found_code');
+const FORBIDDEN_ERROR = require('../errors/forbidden_error');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res
-          .status(ERROR_CODE)
-          .send({ message: 'Переданы некорректные данные при создании пользователя.' });
+        return next(new ERROR_CODE('Переданы некорректные данные при создании пользователя.'));
       }
-      return res
-        .status(INTERNAL_ERROR)
-        .send({ message: 'На сервере произошла ошибка.' });
+      return next();
     });
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
@@ -26,39 +24,33 @@ module.exports.createCard = (req, res) => {
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res
-          .status(ERROR_CODE)
-          .send({ message: 'Переданы некорректные данные при создании пользователя.' });
+        return next(new ERROR_CODE('Переданы некорректные данные при создании пользователя.'));
       }
-      return res
-        .status(INTERNAL_ERROR)
-        .send({ message: 'На сервере произошла ошибка.' });
+      return next();
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        return res
-          .status(ERROR_NOT_FOUND_CODE)
-          .send({ message: 'Передан несуществующий _id карточки.' });
+        return next(new ERROR_NOT_FOUND_CODE('Передан несуществующий _id карточки.'));
       }
-      return res.send({ message: 'Карточка удалена' });
+      if (card.owner.toString() === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => res.status(200));
+      }
+      return next(new FORBIDDEN_ERROR('Удалить чужую карточку нельзя.'));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        return res
-          .status(ERROR_CODE)
-          .send({ message: 'Карточка с указанным _id не найдена.' });
+        return next(new ERROR_CODE('Карточка с указанным _id не найдена.'));
       }
-      return res
-        .status(INTERNAL_ERROR)
-        .send({ message: 'На сервере произошла ошибка.' });
+      return next();
     });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -66,25 +58,19 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        return res
-          .status(ERROR_NOT_FOUND_CODE)
-          .send({ message: 'Передан несуществующий _id карточки.' });
+        return next(new ERROR_NOT_FOUND_CODE('Передан несуществующий _id карточки.'));
       }
       return res.send({ card });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        return res
-          .status(ERROR_CODE)
-          .send({ message: 'Переданы некорректные данные для постановки лайка.' });
+        return next(new ERROR_CODE('Переданы некорректные данные для постановки лайка.'));
       }
-      return res
-        .status(INTERNAL_ERROR)
-        .send({ message: 'На сервере произошла ошибка.' });
+      return next();
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -92,20 +78,14 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        return res
-          .status(ERROR_NOT_FOUND_CODE)
-          .send({ message: 'Передан несуществующий _id карточки.' });
+        return next(new ERROR_NOT_FOUND_CODE('Передан несуществующий _id карточки.'));
       }
       return res.send({ card });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        return res
-          .status(ERROR_CODE)
-          .send({ message: 'Переданы некорректные данные для снятия лайка.' });
+        return next(new ERROR_CODE('Переданы некорректные данные для снятия лайка.'));
       }
-      return res
-        .status(INTERNAL_ERROR)
-        .send({ message: 'На сервере произошла ошибка.' });
+      return next();
     });
 };
